update : divide header & function in C_team4_client
